@@ -18,25 +18,25 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   cloud_firestore: ^3.1.8
   cupertino_icons: ^1.0.2
   date_time_picker: ^2.0.0
   firebase_auth: ^3.3.7
-  firebase_core: ^1.12.0
+  firebase_core: ^1.13.1
   firebase_performance: ^0.8.0+3
-  flutter_icons: ^1.1.0
   flutter:
     sdk: flutter
+  flutter_icons: ^1.1.0
+  flutter_markdown: ^0.6.9
   flutter_svg: ^1.0.3
   flutter_switch: ^0.3.0
   google_fonts: ^2.3.1
   google_sign_in: ^5.2.3
   page_transition: ^2.0.4
   syncfusion_flutter_datepicker: null
-  flutter_markdown: ^0.6.9
 
 dev_dependencies:
   flutter_launcher_icons: "^0.9.0"
