@@ -5,8 +5,10 @@ import 'package:flutter/material.dart';
 import 'package:hotelmanagement/drawer.dart';
 import 'package:hotelmanagement/components/input.dart';
 
+import '../responsive/pageScaffol.dart';
+
 class AggiungiSpeseScreen extends StatefulWidget {
-  const AggiungiSpeseScreen({Key key}) : super(key: key);
+  AggiungiSpeseScreen({Key? key}) : super(key: key);
 
   @override
   _AggiungiSpeseScreenState createState() => _AggiungiSpeseScreenState();
@@ -26,37 +28,37 @@ class _AggiungiSpeseScreenState extends State<AggiungiSpeseScreen> {
   var costoSpesaControlle = TextEditingController();
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(),
-      drawer: const DraweNavigation(),
-      body: Center(
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              const Text(
-                "Aggiungi Spese",
-                style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
+    return PageScaffold(
+        title: "",
+        body: Scaffold(
+          body: Center(
+            child: SingleChildScrollView(
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: [
+                  Text(
+                    "Aggiungi Spese",
+                    style: TextStyle(fontSize: 30, fontWeight: FontWeight.bold),
+                  ),
+                  inputText(TextInputType.text, "Nome spesa", false,
+                      nomeSpesaControlle),
+                  inputText(TextInputType.text, "Descrizione", false,
+                      descrizioneSpesaControlle),
+                  inputText(TextInputType.number, "Costo spesa", false,
+                      costoSpesaControlle),
+                  IconButton(
+                      icon: Icon(Icons.add_shopping_cart),
+                      onPressed: () {
+                        if (nomeSpesaControlle.text.isNotEmpty &&
+                            costoSpesaControlle.text.isNotEmpty) {
+                          addSpesa();
+                        }
+                      }),
+                ],
               ),
-              inputText(
-                  TextInputType.text, "Nome spesa", false, nomeSpesaControlle),
-              inputText(TextInputType.text, "Descrizione", false,
-                  descrizioneSpesaControlle),
-              inputText(TextInputType.number, "Costo spesa", false,
-                  costoSpesaControlle),
-              IconButton(
-                  icon: const Icon(Icons.add_shopping_cart),
-                  onPressed: () {
-                    if (nomeSpesaControlle.text.isNotEmpty &&
-                        costoSpesaControlle.text.isNotEmpty) {
-                      addSpesa();
-                    }
-                  }),
-            ],
+            ),
           ),
-        ),
-      ),
-    );
+        ));
   }
 }
