@@ -4,149 +4,160 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:hotelmanagement/drawer.dart';
+import 'package:hotelmanagement/screen/responsive/pageScaffol.dart';
 import 'package:hotelmanagement/screen/responsive/responsive.dart';
 
 import 'elenco_ospiti.dart';
 
 class ElencoOspitiGenerali extends StatefulWidget {
-  ElencoOspitiGenerali({Key key}) : super(key: key);
+  ElencoOspitiGenerali({Key? key}) : super(key: key);
   @override
   _ElencoOspitiGeneraliState createState() => _ElencoOspitiGeneraliState();
 }
 
 class _ElencoOspitiGeneraliState extends State<ElencoOspitiGenerali> {
-  String cognomePrenotazione;
+  String cognomePrenotazione = "";
 
-  User user = FirebaseAuth.instance.currentUser;
+  User? user = FirebaseAuth.instance.currentUser;
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-          title: const Text(
-        "Hotel Management",
-      )),
-      drawer: const DraweNavigation(),
-      resizeToAvoidBottomInset: false,
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          StreamBuilder(
-              stream: FirebaseFirestore.instance
-                  .collection('users')
-                  .doc(user.uid)
-                  .collection("prenotazioni")
-                  .snapshots(),
-              builder: (context, snapshots) {
-                if (snapshots.hasData) {
-                  return ListView.builder(
-                      shrinkWrap: true,
-                      itemCount: snapshots.data.docs.length,
-                      itemBuilder: (context, index) {
-                        DocumentSnapshot documentSnapshot =
-                            snapshots.data.docs[index];
-                        return Card(
-                            elevation: 4,
-                            margin: const EdgeInsets.all(8),
-                            shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(8)),
-                            child: Column(children: [
-                              ListTile(
-                                title: Padding(
-                                  padding: const EdgeInsets.all(8.0),
-                                  child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Text("Nome Prenotazione: " +
-                                          documentSnapshot["NomePrenotazione"]),
-                                      Text("Cognome Prenotazione: " +
-                                          documentSnapshot[
-                                              "CognomePrenotazione"]),
-                                    ],
-                                  ),
-                                ),
-                                subtitle:
-                                    Text("Piano: " + documentSnapshot["Piano"]),
-                                trailing: IconButton(
-                                  icon: const Icon(Icons.person),
-                                  onPressed: () {
-                                    Navigator.of(context)
-                                        .push(MaterialPageRoute(
-                                            builder: (context) => ElencoOspiti(
-                                                  cognomePrenotazione:
-                                                      documentSnapshot[
-                                                              "CognomePrenotazione"]
-                                                          .toString(),
-                                                )));
-                                  },
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(5.0),
-                                child: Row(
-                                  children: [
-                                    Text("Data di inizio soggiorno: " +
-                                        documentSnapshot["DataDiInizio"]
-                                            .toString()),
-                                  ],
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(5.0),
-                                child: Row(
-                                  children: [
-                                    Text("Data di Fine soggiorno: " +
-                                        documentSnapshot["DataFine"]
-                                            .toString()),
-                                  ],
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(5.0),
-                                child: Row(
-                                  children: [
-                                    Text("Nummero di persone: " +
-                                        documentSnapshot["NPersone"]
-                                            .toString()),
-                                  ],
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(5.0),
-                                child: Row(
-                                  children: [
-                                    Text(
-                                      "Prezzo Soggiorno: " +
-                                          documentSnapshot["Prezzo"]
-                                              .toString() +
-                                          "€",
+    return PageScaffold(
+      title: "Hotel Management",
+      body: Scaffold(
+        resizeToAvoidBottomInset: false,
+        body: SingleChildScrollView(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              StreamBuilder<QuerySnapshot>(
+                  stream: FirebaseFirestore.instance
+                      .collection('users')
+                      .doc(user?.uid)
+                      .collection("prenotazioni")
+                      .snapshots(),
+                  builder: (context, snapshots) {
+                    if (snapshots.hasData) {
+                      return ListView.builder(
+                          shrinkWrap: true,
+                          itemCount: snapshots.data!.docs.length,
+                          itemBuilder: (context, index) {
+                            DocumentSnapshot documentSnapshot =
+                                snapshots.data!.docs[index];
+                            return Card(
+                                elevation: 4,
+                                margin: EdgeInsets.all(8),
+                                shape: RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(8)),
+                                child: Column(children: [
+                                  ListTile(
+                                    title: Padding(
+                                      padding: EdgeInsets.all(8.0),
+                                      child: Column(
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.start,
+                                        children: [
+                                          Text("Nome e cognome Prenotazione: " +
+                                              documentSnapshot[
+                                                  "NomePrenotazione"] +
+                                              " " +
+                                              documentSnapshot[
+                                                  "CognomePrenotazione"]),
+                                        ],
+                                      ),
                                     ),
-                                  ],
-                                ),
-                              ),
-                              Padding(
-                                padding: const EdgeInsets.all(15.0),
-                                child: IconButton(
-                                  icon: const Icon(
-                                    Icons.delete,
-                                    color: Colors.red,
-                                    size: 25,
+                                    subtitle: Padding(
+                                      padding: const EdgeInsets.all(5.0),
+                                      child: Row(
+                                        children: [
+                                          Text("Piano: " +
+                                              documentSnapshot["Piano"]
+                                                  .toString()),
+                                        ],
+                                      ),
+                                    ),
+                                    trailing: IconButton(
+                                      icon: Icon(Icons.person),
+                                      onPressed: () {
+                                        Navigator.of(context).push(
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    ElencoOspiti(
+                                                      cognomePrenotazione:
+                                                          documentSnapshot[
+                                                                  "CognomePrenotazione"]
+                                                              .toString(),
+                                                    )));
+                                      },
+                                    ),
                                   ),
-                                  onPressed: () {},
-                                ),
-                              ),
-                            ]));
-                      });
-                } else {
-                  print("dati non trovati");
-                  return const Align(
-                    alignment: FractionalOffset.center,
-                    child: CircularProgressIndicator(),
-                  );
-                }
-              }),
-        ],
+                                  Padding(
+                                    padding: EdgeInsets.all(5.0),
+                                    child: Row(
+                                      children: [
+                                        Text("Data di inizio soggiorno: " +
+                                            documentSnapshot["DataDiInizio"]
+                                                .toString()),
+                                      ],
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsets.all(5.0),
+                                    child: Row(
+                                      children: [
+                                        Text("Data di Fine soggiorno: " +
+                                            documentSnapshot["DataFine"]
+                                                .toString()),
+                                      ],
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsets.all(5.0),
+                                    child: Row(
+                                      children: [
+                                        Text("Nummero di persone: " +
+                                            documentSnapshot["NPersone"]
+                                                .toString()),
+                                      ],
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsets.all(5.0),
+                                    child: Row(
+                                      children: [
+                                        Text(
+                                          "Prezzo Soggiorno: " +
+                                              documentSnapshot["Prezzo"]
+                                                  .toString() +
+                                              "€",
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                  Padding(
+                                    padding: EdgeInsets.all(15.0),
+                                    child: IconButton(
+                                      icon: Icon(
+                                        Icons.delete,
+                                        color: Colors.red,
+                                        size: 25,
+                                      ),
+                                      onPressed: () {},
+                                    ),
+                                  ),
+                                ]));
+                          });
+                    } else {
+                      print("dati non trovati");
+                      return Align(
+                        alignment: FractionalOffset.center,
+                        child: CircularProgressIndicator(),
+                      );
+                    }
+                  }),
+            ],
+          ),
+        ),
       ),
     );
   }
