@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
 
-// ignore: non_constant_identifier_names
+// ignore: non_ant_identifier_names
 inputText(Textinput, label, obscureText, [var controller]) {
   return Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
     Padding(
-      padding: const EdgeInsets.all(10),
+      padding: EdgeInsets.all(10),
       child: SizedBox(
         width: 500,
         child: TextField(
           decoration: InputDecoration(
-            // ignore: prefer_const_constructors
+            // ignore: prefer__ructors
             border: OutlineInputBorder(
-                borderRadius: const BorderRadius.all(Radius.circular(10))),
+                borderRadius: BorderRadius.all(Radius.circular(10))),
             labelText: label,
             hintText: "inserisci " + label,
           ),
